Update utility function requirements Require non-negative utility functions to prevent errors

+++ b2dp/benches/benchmarks/outcomespace_size_benchmark.rs
@@ -14,7 +14,7 @@ fn run_mechanism(n: i64, optimize: bool) -> u32 {
         outcomes.push(i as u32);
     }
     let options = ExponentialOptions { min_retries: 1, optimized_sample: optimize, empirical_precision: false};
-    let result = exponential_mechanism(eta, &outcomes, utility_fn, 0, n as i64, n as u32, rng, options).unwrap();
+    let result = exponential_mechanism(eta, &outcomes, utility_fn, 0, n as u32, n as u32, rng, options).unwrap();
     *result
 }
