+++ b2dp/benches/benchmarks/timingchannel_benchmark.rs
@@ -18,7 +18,7 @@ fn run_mechanism(num_retries: u32, weight_low: bool) -> u32 {
         outcomes.push(1);
     }
     let options = ExponentialOptions { min_retries: num_retries, optimized_sample: optimize, empirical_precision: false};
-    let result = exponential_mechanism(eta, &outcomes, utility_fn, 0, n as i64, n as u32, rng, options).unwrap();
+    let result = exponential_mechanism(eta, &outcomes, utility_fn, 0, n as u32, n as u32, rng, options).unwrap();
     *result
 }
